requirements.txt
-i https://nexus.alm.europe.cloudcenter.corp/repository/pypi-public/simple
abnf==2.3.1; python_version >= '3.9'
aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.11.16; python_version >= '3.9'
aiosignal==1.3.2; python_version >= '3.9'
anyio==4.9.0; python_version >= '3.9'
attrs==25.3.0; python_version >= '3.8'
bson==0.5.10
cachetools==5.5.2; python_version >= '3.7'
certifi==2025.1.31; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.4.1; python_version >= '3.7'
circuitbreaker==2.1.3
click==8.1.8; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
confluent-kafka==2.9.0; python_version >= '3.7'
coverage[toml]==7.6.12; python_version >= '3.9'
cryptography==44.0.2; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
darwin-composer==4.2.41729588485684; python_version >= '3.9'
darwin-error-handler==4.1.21730210948533; python_version >= '3.9'
darwin-logging==4.2.3; python_version >= '3.9'
darwin-middlewares==4.1.21732129575032; python_version >= '3.9'
darwin-security==4.2.21729523967095; python_version >= '3.9'
deprecated==1.2.18; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
dnspython==2.7.0; python_version >= '3.9'
durationpy==0.9
ecs-logging==2.1.0; python_version >= '3.6'
fastapi==0.115.12; python_version >= '3.8'
fastapi-health==0.4.0; python_full_version >= '3.6.1' and python_full_version < '4.0.0'
frozenlist==1.5.0; python_version >= '3.8'
gevent==24.11.1; python_version >= '3.9'
google-auth==2.38.0; python_version >= '3.7'
greenlet==3.1.1; python_version >= '3.7'
gunicorn==23.0.0; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httptools==0.6.4; python_full_version >= '3.8.0'
httpx==0.27.2; python_version >= '3.8'
hvac==2.3.0; python_version >= '3.8' and python_version < '4.0'
idna==3.10; python_version >= '3.6'
importlib-metadata==8.6.1; python_version >= '3.9'
iniconfig==2.1.0; python_version >= '3.8'
kubernetes==32.0.1; python_version >= '3.6'
memoization==0.4.0; python_version >= '3' and python_version not in '3.0, 3.1, 3.2, 3.3' and python_version < '4'
multidict==6.2.0; python_version >= '3.9'
oauthlib==3.2.2; python_version >= '3.6'
opentelemetry-api==1.31.1; python_version >= '3.8'
opentelemetry-propagator-b3==1.31.1; python_version >= '3.8'
opentelemetry-sdk==1.31.1; python_version >= '3.8'
opentelemetry-semantic-conventions==0.52b1; python_version >= '3.8'
packaging==24.2; python_version >= '3.8'
pip==25.0.1; python_version >= '3.8'
pipdeptree==2.26.0; python_version >= '3.9'
pluggy==1.5.0; python_version >= '3.8'
propcache==0.3.1; python_version >= '3.9'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pydantic==1.10.7; python_version >= '3.7'
pydash==8.0.5; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
pymongo==4.6.3; python_version >= '3.7'
pytest==8.3.5; python_version >= '3.8'
pytest-asyncio==0.26.0; python_version >= '3.9'
pytest-cov==6.1.0; python_version >= '3.9'
pytest-env==1.1.5; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.1.0; python_version >= '3.9'
python-i18n==0.3.9
pytz==2023.3
pyyaml==6.0.2; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rsa==4.9; python_version >= '3.6' and python_version < '4'
setuptools==78.1.0; python_version >= '3.9'
shuttlelib==2.0.21732883155486; python_version >= '3.9'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.46.1; python_version >= '3.9'
typing-extensions==4.13.0; python_version >= '3.8'
urllib3==1.26.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
uvicorn[standard]==0.29.0; python_version >= '3.8'
validators==0.34.0; python_version >= '3.8'
watchfiles==1.0.4; python_version >= '3.9'
websocket==0.2.1
websocket-client==1.8.0; python_version >= '3.8'
websockets==15.0.1; python_version >= '3.9'
wrapt==1.17.2; python_version >= '3.8'
yarl==1.18.3; python_version >= '3.9'
zipp==3.21.0; python_version >= '3.9'
zope.event==5.0; python_version >= '3.7'
zope.interface==7.2; python_version >= '3.8'
